# backend/internal/pkg/api/photo_service_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.1
// source: photo_service.proto

package api

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationPhotoServiceDeletePhoto = "/api.PhotoService/DeletePhoto"
const OperationPhotoServiceGetPhotoDetail = "/api.PhotoService/GetPhotoDetail"
const OperationPhotoServiceGetPhotoStats = "/api.PhotoService/GetPhotoStats"
const OperationPhotoServiceSearchPhotos = "/api.PhotoService/SearchPhotos"
const OperationPhotoServiceUpdatePhotoTags = "/api.PhotoService/UpdatePhotoTags"

type PhotoServiceHTTPServer interface {
	// DeletePhoto 删除照片（含文件清理）
	DeletePhoto(context.Context, *DeletePhotoRequest) (*DeletePhotoResponse, error)
	// GetPhotoDetail 单张详情
	GetPhotoDetail(context.Context, *GetPhotoDetailRequest) (*GetPhotoDetailResponse, error)
	// GetPhotoStats 综合统计
	GetPhotoStats(context.Context, *Empty) (*GetPhotoStatsResponse, error)
	// SearchPhotos 复杂条件分页查询
	SearchPhotos(context.Context, *SearchPhotosRequest) (*SearchPhotosResponse, error)
	// UpdatePhotoTags 更新标签
	UpdatePhotoTags(context.Context, *UpdatePhotoTagsRequest) (*Empty, error)
}

func RegisterPhotoServiceHTTPServer(s *http.Server, srv PhotoServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/photos", _PhotoService_SearchPhotos0_HTTP_Handler(srv))
	r.GET("/api/v1/photos/stats", _PhotoService_GetPhotoStats0_HTTP_Handler(srv))
	r.GET("/api/v1/photos/{id}", _PhotoService_GetPhotoDetail0_HTTP_Handler(srv))
	r.PUT("/api/v1/photos/{id}/tags", _PhotoService_UpdatePhotoTags0_HTTP_Handler(srv))
	r.DELETE("/api/v1/photos/{id}", _PhotoService_DeletePhoto0_HTTP_Handler(srv))
}

func _PhotoService_SearchPhotos0_HTTP_Handler(srv PhotoServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SearchPhotosRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPhotoServiceSearchPhotos)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SearchPhotos(ctx, req.(*SearchPhotosRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SearchPhotosResponse)
		return ctx.Result(200, reply)
	}
}

func _PhotoService_GetPhotoStats0_HTTP_Handler(srv PhotoServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPhotoServiceGetPhotoStats)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetPhotoStats(ctx, req.(*Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetPhotoStatsResponse)
		return ctx.Result(200, reply)
	}
}

func _PhotoService_GetPhotoDetail0_HTTP_Handler(srv PhotoServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetPhotoDetailRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPhotoServiceGetPhotoDetail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetPhotoDetail(ctx, req.(*GetPhotoDetailRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetPhotoDetailResponse)
		return ctx.Result(200, reply)
	}
}

func _PhotoService_UpdatePhotoTags0_HTTP_Handler(srv PhotoServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdatePhotoTagsRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPhotoServiceUpdatePhotoTags)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdatePhotoTags(ctx, req.(*UpdatePhotoTagsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Empty)
		return ctx.Result(200, reply)
	}
}

func _PhotoService_DeletePhoto0_HTTP_Handler(srv PhotoServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeletePhotoRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPhotoServiceDeletePhoto)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeletePhoto(ctx, req.(*DeletePhotoRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeletePhotoResponse)
		return ctx.Result(200, reply)
	}
}

type PhotoServiceHTTPClient interface {
	// DeletePhoto 删除照片（含文件清理）
	DeletePhoto(ctx context.Context, req *DeletePhotoRequest, opts ...http.CallOption) (rsp *DeletePhotoResponse, err error)
	// GetPhotoDetail 单张详情
	GetPhotoDetail(ctx context.Context, req *GetPhotoDetailRequest, opts ...http.CallOption) (rsp *GetPhotoDetailResponse, err error)
	// GetPhotoStats 综合统计
	GetPhotoStats(ctx context.Context, req *Empty, opts ...http.CallOption) (rsp *GetPhotoStatsResponse, err error)
	// SearchPhotos 复杂条件分页查询
	SearchPhotos(ctx context.Context, req *SearchPhotosRequest, opts ...http.CallOption) (rsp *SearchPhotosResponse, err error)
	// UpdatePhotoTags 更新标签
	UpdatePhotoTags(ctx context.Context, req *UpdatePhotoTagsRequest, opts ...http.CallOption) (rsp *Empty, err error)
}

type PhotoServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewPhotoServiceHTTPClient(client *http.Client) PhotoServiceHTTPClient {
	return &PhotoServiceHTTPClientImpl{client}
}

// DeletePhoto 删除照片（含文件清理）
func (c *PhotoServiceHTTPClientImpl) DeletePhoto(ctx context.Context, in *DeletePhotoRequest, opts ...http.CallOption) (*DeletePhotoResponse, error) {
	var out DeletePhotoResponse
	pattern := "/api/v1/photos/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPhotoServiceDeletePhoto))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetPhotoDetail 单张详情
func (c *PhotoServiceHTTPClientImpl) GetPhotoDetail(ctx context.Context, in *GetPhotoDetailRequest, opts ...http.CallOption) (*GetPhotoDetailResponse, error) {
	var out GetPhotoDetailResponse
	pattern := "/api/v1/photos/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPhotoServiceGetPhotoDetail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetPhotoStats 综合统计
func (c *PhotoServiceHTTPClientImpl) GetPhotoStats(ctx context.Context, in *Empty, opts ...http.CallOption) (*GetPhotoStatsResponse, error) {
	var out GetPhotoStatsResponse
	pattern := "/api/v1/photos/stats"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPhotoServiceGetPhotoStats))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SearchPhotos 复杂条件分页查询
func (c *PhotoServiceHTTPClientImpl) SearchPhotos(ctx context.Context, in *SearchPhotosRequest, opts ...http.CallOption) (*SearchPhotosResponse, error) {
	var out SearchPhotosResponse
	pattern := "/api/v1/photos"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPhotoServiceSearchPhotos))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdatePhotoTags 更新标签
func (c *PhotoServiceHTTPClientImpl) UpdatePhotoTags(ctx context.Context, in *UpdatePhotoTagsRequest, opts ...http.CallOption) (*Empty, error) {
	var out Empty
	pattern := "/api/v1/photos/{id}/tags"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPhotoServiceUpdatePhotoTags))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}