# backend/internal/pkg/api/query_service_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.1
// source: query_service.proto

package api

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationQueryServiceExecuteSQL = "/api.QueryService/ExecuteSQL"
const OperationQueryServiceGetAttributeValues = "/api.QueryService/GetAttributeValues"
const OperationQueryServiceGetPhotoSchema = "/api.QueryService/GetPhotoSchema"

type QueryServiceHTTPServer interface {
	// ExecuteSQL 执行 SELECT SQL 查询
	ExecuteSQL(context.Context, *ExecuteSQLRequest) (*ExecuteSQLResponse, error)
	// GetAttributeValues 返回所有结构化属性的去重值（供 Text-to-SQL prompt 拼接）
	GetAttributeValues(context.Context, *Empty) (*GetAttributeValuesResponse, error)
	// GetPhotoSchema 返回 photos 表结构
	GetPhotoSchema(context.Context, *Empty) (*GetPhotoSchemaResponse, error)
}

func RegisterQueryServiceHTTPServer(s *http.Server, srv QueryServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/sql/query", _QueryService_ExecuteSQL0_HTTP_Handler(srv))
	r.GET("/api/v1/sql/photos/schema", _QueryService_GetPhotoSchema0_HTTP_Handler(srv))
	r.GET("/api/v1/sql/photos/attribute-values", _QueryService_GetAttributeValues0_HTTP_Handler(srv))
}

func _QueryService_ExecuteSQL0_HTTP_Handler(srv QueryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ExecuteSQLRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationQueryServiceExecuteSQL)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ExecuteSQL(ctx, req.(*ExecuteSQLRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ExecuteSQLResponse)
		return ctx.Result(200, reply)
	}
}

func _QueryService_GetPhotoSchema0_HTTP_Handler(srv QueryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationQueryServiceGetPhotoSchema)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetPhotoSchema(ctx, req.(*Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetPhotoSchemaResponse)
		return ctx.Result(200, reply)
	}
}

func _QueryService_GetAttributeValues0_HTTP_Handler(srv QueryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationQueryServiceGetAttributeValues)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetAttributeValues(ctx, req.(*Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetAttributeValuesResponse)
		return ctx.Result(200, reply)
	}
}

type QueryServiceHTTPClient interface {
	// ExecuteSQL 执行 SELECT SQL 查询
	ExecuteSQL(ctx context.Context, req *ExecuteSQLRequest, opts ...http.CallOption) (rsp *ExecuteSQLResponse, err error)
	// GetAttributeValues 返回所有结构化属性的去重值（供 Text-to-SQL prompt 拼接）
	GetAttributeValues(ctx context.Context, req *Empty, opts ...http.CallOption) (rsp *GetAttributeValuesResponse, err error)
	// GetPhotoSchema 返回 photos 表结构
	GetPhotoSchema(ctx context.Context, req *Empty, opts ...http.CallOption) (rsp *GetPhotoSchemaResponse, err error)
}

type QueryServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewQueryServiceHTTPClient(client *http.Client) QueryServiceHTTPClient {
	return &QueryServiceHTTPClientImpl{client}
}

// ExecuteSQL 执行 SELECT SQL 查询
func (c *QueryServiceHTTPClientImpl) ExecuteSQL(ctx context.Context, in *ExecuteSQLRequest, opts ...http.CallOption) (*ExecuteSQLResponse, error) {
	var out ExecuteSQLResponse
	pattern := "/api/v1/sql/query"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationQueryServiceExecuteSQL))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetAttributeValues 返回所有结构化属性的去重值（供 Text-to-SQL prompt 拼接）
func (c *QueryServiceHTTPClientImpl) GetAttributeValues(ctx context.Context, in *Empty, opts ...http.CallOption) (*GetAttributeValuesResponse, error) {
	var out GetAttributeValuesResponse
	pattern := "/api/v1/sql/photos/attribute-values"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationQueryServiceGetAttributeValues))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetPhotoSchema 返回 photos 表结构
func (c *QueryServiceHTTPClientImpl) GetPhotoSchema(ctx context.Context, in *Empty, opts ...http.CallOption) (*GetPhotoSchemaResponse, error) {
	var out GetPhotoSchemaResponse
	pattern := "/api/v1/sql/photos/schema"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationQueryServiceGetPhotoSchema))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}