# backend/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/photos:
        get:
            tags:
                - PhotoService
            description: 复杂条件分页查询
            operationId: PhotoService_SearchPhotos
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: timeline
                  in: query
                  schema:
                    type: string
                - name: tag
                  in: query
                  schema:
                    type: string
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: brand
                  in: query
                  schema:
                    type: string
                - name: lens
                  in: query
                  schema:
                    type: string
                - name: focalMin
                  in: query
                  schema:
                    type: string
                - name: focalMax
                  in: query
                  schema:
                    type: string
                - name: isoMin
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: isoMax
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: shotAtStart
                  in: query
                  schema:
                    type: string
                - name: shotAtEnd
                  in: query
                  schema:
                    type: string
                - name: sortBy
                  in: query
                  schema:
                    type: string
                - name: sortOrder
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.SearchPhotosResponse'
...
    /api/v1/photos/{id}:
        get:
            tags:
                - PhotoService
            description: 单张详情
            operationId: PhotoService_GetPhotoDetail
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.GetPhotoDetailResponse'
        delete:
            tags:
                - PhotoService
            description: 删除照片（含文件清理）
            operationId: PhotoService_DeletePhoto
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.DeletePhotoResponse'
...
    /api/v1/photos/{id}/tags:
        put:
            tags:
                - PhotoService
            description: 更新标签
            operationId: PhotoService_UpdatePhotoTags
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.UpdatePhotoTagsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.Empty'
...
    /api/v1/sql/query:
        post:
            tags:
                - QueryService
            description: 执行 SELECT SQL 查询
            operationId: QueryService_ExecuteSQL
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.ExecuteSQLRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ExecuteSQLResponse'
...
    /api/v1/tags/{name}/photos:
        get:
            tags:
                - TagService
            description: 某标签下的照片
            operationId: TagService_GetPhotosByTag
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.GetPhotosByTagResponse'
...
    /api/v1/timelines/{name}/photos:
        get:
            tags:
                - TimelineService
            description: 某时间线下的照片
            operationId: TimelineService_GetPhotosByTimeline
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.GetPhotosByTimelineResponse'
...
    /photos:
        get:
            tags:
                - DefaultCURD
            operationId: DefaultCURD_GetPhotoList
            parameters:
                - name: IDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.GetPhotoListResponse'
        post:
            tags:
                - DefaultCURD
            description: |-
                --------------------------------------------------
                 tbl : photos
            operationId: DefaultCURD_AddPhoto
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.AddPhotoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.AddPhotoResponse'
        delete:
            tags:
                - DefaultCURD
            operationId: DefaultCURD_DelPhotoByIDList
            parameters:
                - name: IDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.Empty'
        patch:
            tags:
                - DefaultCURD
            operationId: DefaultCURD_UpdatePhoto
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.UpdatePhotoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.UpdatePhotoResponse'
components:
    schemas:
...
        api.AddPhotoRequest:
            type: object
            properties:
                photo:
                    $ref: '#/components/schemas/api.PhotoInfo'
        api.AddPhotoResponse:
            type: object
            properties:
                photo:
                    $ref: '#/components/schemas/api.PhotoInfo'
...
        api.DeletePhotoResponse:
            type: object
            properties:
                status:
                    type: string
                id:
                    type: string
...
        api.Empty:
            type: object
            properties: {}
        api.ExecuteSQLRequest:
            type: object
            properties:
                sql:
                    type: string
                limit:
                    type: integer
                    format: int32
        api.ExecuteSQLResponse:
            type: object
            properties:
                columns:
                    type: array
                    items:
                        type: string
                rows:
                    type: array
                    items:
                        type: object
                count:
                    type: integer
                    format: int32
...
        api.GetPhotoDetailResponse:
            type: object
            properties:
                photo:
                    $ref: '#/components/schemas/api.PhotoItem'
                imageUrl:
                    type: string
                descriptionModel:
                    type: string
                descriptionTime:
                    type: string
        api.GetPhotoListResponse:
            type: object
            properties:
                photoList:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.PhotoInfo'
...
        api.GetPhotosByTagResponse:
            type: object
            properties:
                tag:
                    type: string
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.PhotoItem'
                total:
                    type: integer
                    format: int32
        api.GetPhotosByTimelineResponse:
            type: object
            properties:
                timeline:
                    type: string
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.PhotoItem'
                total:
                    type: integer
                    format: int32
...
        api.PhotoInfo:
            type: object
            properties:
                ID:
                    type: string
                filename:
                    type: string
                filePath:
                    type: string
                timeline:
                    type: string
                tags:
                    type: string
                description:
                    type: string
                objects:
                    type: string
                colors:
                    type: string
                scene:
                    type: string
                lighting:
                    type: string
                mood:
                    type: string
                composition:
                    type: string
                shotAt:
                    type: string
                width:
                    type: integer
                    format: int32
                height:
                    type: integer
                    format: int32
                brand:
                    type: string
                model:
                    type: string
                lens:
                    type: string
                focalLength:
                    type: string
                aperture:
                    type: string
                iso:
                    type: integer
                    format: int32
                exposureTime:
                    type: string
                latitude:
                    type: number
                    format: double
                longitude:
                    type: number
                    format: double
                altitude:
                    type: number
                    format: double
                importedAt:
                    type: string
            description: |-
                --------------------------------------------------
                 tbl : photos
        api.PhotoItem:
            type: object
            properties:
                id:
                    type: string
                filename:
                    type: string
                filePath:
                    type: string
                timeline:
                    type: string
                tags:
                    type: string
                description:
                    type: string
                objects:
                    type: string
                colors:
                    type: string
                scene:
                    type: string
                lighting:
                    type: string
                mood:
                    type: string
                composition:
                    type: string
                shotAt:
                    type: string
                width:
                    type: integer
                    format: int32
                height:
                    type: integer
                    format: int32
                brand:
                    type: string
                model:
                    type: string
                lens:
                    type: string
                focalLength:
                    type: string
                aperture:
                    type: string
                iso:
                    type: integer
                    format: int32
                exposureTime:
                    type: string
                latitude:
                    type: number
                    format: double
                longitude:
                    type: number
                    format: double
                altitude:
                    type: number
                    format: double
                importedAt:
                    type: string
                hasDescription:
                    type: boolean
                thumbnailUrl:
                    type: string
...
        api.SearchPhotosResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.PhotoItem'
                total:
                    type: string
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                totalPages:
                    type: integer
                    format: int32
...
        api.UpdatePhotoRequest:
            type: object
            properties:
                photo:
                    $ref: '#/components/schemas/api.PhotoInfo'
        api.UpdatePhotoResponse:
            type: object
            properties:
                photo:
                    $ref: '#/components/schemas/api.PhotoInfo'
        api.UpdatePhotoTagsRequest:
            type: object
            properties:
                id:
                    type: string
                tags:
                    type: string
...
servers:
    - description: PGO API
      url: http://127.0.0.1:8080